# docs/api/openapi.json
{
  "components": {
...
    "schemas": {
...
      "PaymentCreateRequest": {
        "properties": {
          "amount": {
            "minLength": 1,
            "type": "string"
          },
          "currency": {
            "minLength": 1,
            "pattern": "[A-Z]{3}",
            "type": "string"
          },
          "paymentMethodId": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency",
          "paymentMethodId"
        ],
        "type": "object"
      },
...
      "PaymentView": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "failureReason": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "paymentMethodId": {
            "type": "string"
          },
          "refundPending": {
            "type": "string"
          },
          "refunded": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RefundRequest": {
        "properties": {
          "amount": {
            "minLength": 1,
            "type": "string"
          },
          "currency": {
            "minLength": 1,
            "pattern": "[A-Z]{3}",
            "type": "string"
          },
          "reason": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency",
          "reason"
        ],
        "type": "object"
      },
      "RefundView": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SessionView": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "checkoutId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "lineSummary": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "paymentIntentId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The public HTTP contract. Versioned in the path (ADR-0015); errors follow RFC 9457, catalogued in docs/architecture/ERROR_CONTRACT.md. Generated from the application on every build and compared against this committed copy, so a change to the contract cannot reach a client without being reviewed.",
    "title": "finapp platform API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/checkout/sessions/{id}": {
      "get": {
        "operationId": "viewCheckoutSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/payments": {
      "get": {
        "operationId": "listPayments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaymentView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "operationId": "createPayment",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
    "/v1/payments/{id}": {
      "delete": {
        "operationId": "cancelPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "get": {
        "operationId": "findPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/payments/{id}/confirmation": {
      "post": {
        "operationId": "confirmPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/payments/{id}/refund": {
      "post": {
        "operationId": "refundPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/v1/sessions/{id}": {
      "delete": {
        "operationId": "revokeSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  }
}